backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
langchain
langchain-community
langchain-groq
llama-index
chromadb
neo4j
transformers
torch
sentence-transformers
xgboost
scikit-learn

groq
supabase
python-multipart
celery
redis
httpx
jinja2
sentry-sdk
langsmith
duckduckgo-search
pytesseract
Pillow
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pandas
openpyxl
